+++ src/main/java/com/kruemblegard/init/ModBlocks.java
@@ -43,6 +43,7 @@ import com.kruemblegard.block.ScarstoneBlock;
 import com.kruemblegard.block.VoidfeltBlock;
 import com.kruemblegard.block.BuoyancyLiftPlatformBlock;
 import com.kruemblegard.block.ConveyorMembraneBlock;
+import com.kruemblegard.block.PressureLoomBlock;
 import com.kruemblegard.block.PaleweftGrassBlock;
 import com.kruemblegard.block.PaleweftCornCropBlock;
 import com.kruemblegard.block.WayfallReactivePlantBlock;
@@ -202,6 +203,15 @@ public final class ModBlocks {
                     .mapColor(MapColor.QUARTZ)
                     .requiresCorrectToolForDrops()
                     .strength(2.2F, 6.0F)
+                    .sound(SoundType.STONE))
+    );
+
+    public static final RegistryObject<Block> PRESSURE_LOOM = BLOCKS.register(
+            "pressure_loom",
+            () -> new PressureLoomBlock(BlockBehaviour.Properties.of()
+                    .mapColor(MapColor.QUARTZ)
+                    .requiresCorrectToolForDrops()
+                    .strength(2.8F, 8.0F)
                     .sound(SoundType.STONE))
     );